+++ eido/__init__.py
@@ -23,4 +23,7 @@ __all__ = [
     "yaml_pep_filter",
     "csv_pep_filter",
     "yaml_samples_pep_filter",
+    "EidoValidationError",
+    "validate_input_files",
+    "get_input_files_size",
 ]