+++ eido/exceptions.py
@@ -2,7 +2,12 @@
 
 from abc import ABCMeta
 
-_all__ = ["PathAttrNotFoundError", "EidoSchemaInvalidError", "EidoFilterError"]
+_all__ = [
+    "EidoFilterError",
+    "EidoSchemaInvalidError",
+    "EidoValidationError",
+    "PathAttrNotFoundError",
+]
 
 
 class EidoException(Exception):
@@ -35,6 +40,6 @@ class EidoFilterError(EidoException):
 class EidoValidationError(EidoException):
     """Object was not validated successfully according to schema."""
 
-    def __init__(self, message, errors):
+    def __init__(self, message, errors_by_type):
         super().__init__(message)
-        self.errors = errors
+        self.errors_by_type = errors_by_type