+++ eido/output_formatters.py
@@ -26,7 +26,6 @@ class MultilineOutputFormatter(BaseOutputFormatter):
         header = MultilineOutputFormatter._get_header(sample_attributes)
 
         for sample in samples:
-
             attribute_with_multiple_properties = MultilineOutputFormatter._get_the_name_of_the_first_attribute_with_multiple_properties(
                 sample, sample_attributes
             )
@@ -103,14 +102,15 @@ class MultilineOutputFormatter(BaseOutputFormatter):
             Representation of sample as a CSV row.
         """
         sample_row = []
-        for attribute in sample_attributes:
 
-            if MultilineOutputFormatter._sample_attribute_is_list(
-                sample, attribute
-            ) and getattr(sample, attribute):
-                value = getattr(sample, attribute)[sample_index]
+        for attribute in sample_attributes:
+            if (
+                MultilineOutputFormatter._sample_attribute_is_list(sample, attribute)
+                and sample[attribute]
+            ):
+                value = sample[attribute][sample_index]
             else:
-                value = getattr(sample, attribute, "")
+                value = sample[attribute]
 
             sample_row.append(value or "")