+++ tests/test_validations.py
@@ -6,7 +6,7 @@ from peppy import Project
 from peppy.utils import load_yaml
 
 from eido import *
-from eido.exceptions import EidoValidationError
+from eido.exceptions import EidoValidationError, PathAttrNotFoundError
 
 
 def _check_remote_file_accessible(url):
@@ -57,7 +57,9 @@ class TestProjectValidation:
 
 class TestSampleValidation:
     @pytest.mark.parametrize("sample_name", [0, 1, "GSM1558746"])
-    def test_validate_works(self, project_object, sample_name, schema_samples_file_path):
+    def test_validate_works(
+        self, project_object, sample_name, schema_samples_file_path
+    ):
         validate_sample(
             project=project_object,
             sample_name=sample_name,
@@ -136,3 +138,11 @@ class TestProjectWithoutConfigValidation:
             validate_project(
                 project=Project(remote_pep_cfg), schema=schema_invalid_file_path
             )
+
+    def test_validate_file_existance(
+        self, test_file_existance_pep, test_file_existence_schema
+    ):
+        schema_path = test_file_existence_schema
+        prj = Project(test_file_existance_pep)
+        with pytest.raises(PathAttrNotFoundError):
+            validate_input_files(prj, schema_path)